Rename BoardButtons and BoardSelectionTab to SpaceButtons and SpaceSelectionTab
