+++ ui/src/Application/PeopleMover.scss
@@ -21,7 +21,7 @@
     overflow: hidden;
 }
 
-.boardButtons button {
+.spaceButtons button {
     margin: 0 12px;
 }
